extract category input component
